Refine CLI help execution flow
Route help requests through stdout with exit code zero, keep real failures on stderr, and split help text by command domain for easier maintenance. Add nested help dispatch so subcommand help paths behave consistently across bundle, node, and site commands.
